Initital upload of solar power cost-benefit analysis
This analysis of the cost-benefit of solar power for my property. It uses data from the NIWA SolarView site, which is superb in terms of allowing accurate estimations of sunhours for actual locations taking account of topography also. Short answer is that there is, at best, a break even of 10 years. This is hardly worth doing, The killer is the cost of storage using a battery. The second barrier is the incredibly low rate power companies buy back from consumers, which renders either storage option pitful really.